fix FixedSizeQueue.Count return type from decimal to int

+++ src/Serilog.Sinks.TCP/Splunk.Logging/FixedSizeQueue.cs
@@ -77,6 +77,6 @@ namespace Splunk.Logging
         }
 
 
-        public decimal Count { get { return _collection.Count; } }
+        public int Count { get { return _collection.Count; } }
     }
 }
